+++ src/main/java/uk/gov/hmcts/register/fees/service/FeesRegisterService.java
@@ -1,18 +1,20 @@
 package uk.gov.hmcts.register.fees.service;
 
-import java.math.BigDecimal;
 import java.util.List;
-
-import uk.gov.hmcts.register.fees.loader.Category;
-import uk.gov.hmcts.register.fees.loader.Fee;
-import uk.gov.hmcts.register.fees.loader.FeesRegister;
+import uk.gov.hmcts.register.fees.model.Category;
+import uk.gov.hmcts.register.fees.model.Fee;
+import uk.gov.hmcts.register.fees.model.FeesRegister;
 
 public interface FeesRegisterService {
-	
-	 FeesRegister getFeesRegister() ;
-	 List<Category> getAllCategories();
-	 Fee getFeeDetails(String eventId);
-	 Fee getFeeDetails(String eventId, BigDecimal claimAmount);
-	 Fee getFeeDetailsForClaimAmountAndCategory(BigDecimal claimAmount, String categoryId);
+
+    FeesRegister getFeesRegister();
+
+    List<Category> getAllCategories();
+
+    Fee getFeeDetails(String eventId);
+
+    Fee getFeeDetails(String eventId, int claimAmount);
+
+    Fee getFeeDetailsForClaimAmountAndCategory(int claimAmount, String categoryId);
 
 }
